+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Don't consider empty files to have any lines
+
 ## [0.1.0] - 2022-02-07
 ### Changed
 - Return error from `Lhdiff`